relinted and merged delete into thread

+++ Forum/Thread/views.py
@@ -32,17 +32,14 @@ def edit_profile(request):
 
 def thread(request):
     if request.POST:
-        p = Post(userid=request.POST['userid'], text=request.POST['text'], time=datetime.datetime.now)
-        p.save()
+        if request.POST['todo'] == 'add':
+            p = Post(userid=request.POST['userid'], text=request.POST['text'], time=datetime.datetime.now)
+            p.save()
+        elif request.POST['todo'] == 'del':
+            id_to_delete = request.POST['del_id']
+            post_to_delete = Post.objects.get(id=id_to_delete)
+            post_to_delete.delete()
     posts = Post.objects.order_by('time')
     form = submit_post()
-    return render_to_response('thread.html', {'posts': posts,'form':form}, context_instance=RequestContext(request))
-
-def delete(request):
-    #assuming the imput is correct (BAD)
-    id_to_delete = request.POST['del_id']
-    post_to_delete = Post.objects.get(id=id_to_delete)
-    post_to_delete.delete()
-
-    #redirect back to thread
+    return render_to_response('thread.html', {'posts': posts, 'form': form}, context_instance=RequestContext(request))
     return HttpResponseRedirect("/forum/")
